+++ lib/jsonapi/resources/matchers.rb
@@ -2,6 +2,7 @@ require "jsonapi/resources/matchers/version"
 require "jsonapi/resources/matchers/have_attribute"
 require "jsonapi/resources/matchers/filter"
 require "jsonapi/resources/matchers/have_many"
+require "jsonapi/resources/matchers/have_one"
 require "jsonapi/resources/matchers/have_model_name"
 require "jsonapi/resources/matchers/have_primary_key"
 
@@ -19,6 +20,10 @@ module JSONAPI
 
       def have_many(name)
         HaveMany.new(name)
+      end
+
+      def have_one(name)
+        HaveOne.new(name)
       end
 
       def have_model_name(name)