+++ spec/dummy/app/resources/libro_resource.rb
@@ -1,5 +1,6 @@
 class LibroResource < JSONAPI::Resource
 
   model_name "Book"
+  has_one :author
 
 end