+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Add `have_one` matcher
+
 # v0.1.0
 
 - Initial release